Implemented DTO Validation, ResponseEntity<> for proper HttpStatus code and throw reasonable custom message, Global and Custom Exception Handling

+++ src/main/java/com/mymart/entity/Category.java
@@ -17,7 +17,7 @@ public class Category
 {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)  
-  private long categoryId;
+  private Integer categoryId;
   @Column(unique=true)
   private String name;
   private String categoryDescription;
